facilities-controllers.js: created to separate route and controller
